old now

+++ controllers/entityController.js
@@ -2,28 +2,9 @@ const { Router } = require('express');
 const router = Router();
 const passport = require('passport');
 
-module.exports = configureRoleRoutes;
-
-function configureRoleRoutes() {
-  router.get(
-    '/api/getRoles',
-    passport.authenticate('jwt', { session: false }),
-    function(req, res) {
-      RoleModel.find()
-        .populate('type')
-        .exec(function(err, roles) {
-          if (err) {
-            res.send('find no good' + err);
-          } else {
-            res.json(roles);
-          }
-        });
-    }
-  );
-}
-
 router.post('/api/addEntity', passport.authenticate('jwt', { session: false }), function (req, res, next) {
-    var entity = getEntityModel(req.query.entityName, next) 
+    var entityModel = getEntityModel(req.body.entityName, next)
+    var entity = new entityModel(req.body.entity)
     entity.save((err, newEntity) => {
         if (err) {
             return next(err);
@@ -46,37 +27,37 @@ router.get('/api/getEntity/:id', passport.authenticate('jwt', { session: false }
         });
 });
 
-router.get('/api/getEntities', passport.authenticate('jwt', { session: false }), function (req, res, next) {
-    var entity = getEntityModel(req.query.entityName, next)
-    entity.find(function (err, entities) {
-        if (err) {
-            next('getEntities no good' + err);
+router.get('/api/getEntities', passport.authenticate('jwt', { session: false }), async function (req, res, next) {
+    var entityName = req.query.entityName
+    if(entityName){
+        var entities = await getEntitiesBy(entityName)
+        if(entities){
+            res.json(entities);            
         }
-        else {
-            res.json(entities);
-        }
-    })
+    }
 });
 
-router.put('/api/updateEntity/:id', passport.authenticate('jwt', { session: false }), function (req, res, next) {
-    console.log('updating entity: ' + req.body.entityData.name)
+router.put('/api/updateEntity', passport.authenticate('jwt', { session: false }), function (req, res, next) {
     var entity = getEntityModel(req.body.entityName, next)
     entity.findOneAndUpdate(
-        { _id: req.params.id },
-        req.body.entityData,
-        { upsert: true },
+        { _id: req.body.entity._id },
+        req.body.entity,
+        { upsert: true , new: true, returnOriginal: false},
         function (err, newEntity) {
             if (err) {
                 next('Error updating Entity\n' + err);
             }
             else {
-                res.send(204);
+                res.json(newEntity);
             }
         });
 });
 
-router.delete('/api/deleteEntity/:id', passport.authenticate('jwt', { session: false }), function (req, res, next) {
-    var entity = getEntityModel(req.body.entityName, next)
+router.delete('/api/deleteEntity/:id/:entityName', passport.authenticate('jwt', { session: false }), function (req, res, next) {
+    if (!req.params.entityName || !req.params.id) {
+        next('Error deleting Entity\n' + err);
+    }
+    var entity = getEntityModel(req.params.entityName, next)
     entity.findOneAndRemove(
         { _id: req.params.id },
         function (err, newEntity) {
@@ -105,16 +86,29 @@ router.get('/api/findEntityLike', passport.authenticate('jwt', { session: false 
     });
 });
 
+async function getEntitiesBy(name, next){
+    var entity = getEntityModel(name, next)
+    return entity.find(function (err, entities) {
+        if (err) {
+            next('getEntities no good' + err)
+        }
+        else {
+            return(entities)
+        }
+    })
+}
+
 function getEntityModel(name, next){
     var entityName = name
     if(!entityName){
         next('Entity Name no good');
     }
     try {
-        var entity = require('./models/' + entityName)
+        var entity = require('../models/' + entityName)
     }
     catch (err) {
         next('Entity path ./models/' + entityName + ' not found\n' + err)
     }
     return entity
 }
+    module.exports = router
